Consolidate action modules
